patch from Rafał Kwaśny mag at entropy.be
Hi,
I encountered a bug while making bootable usb drive with persistent
storage and using clicfs to do so.

If the time on the appliance is set in the future ( for ex due to broken
hwclock ) and after startup time is synced via NTP clicfs stops writing
changes to disk. If the system is rebooted all changes are lost.

Please see attached patch that changes time() to uptime for syncing
delay. It seems to work but it's not throughly tested.